+++ src/main/java/ovh/heraud/nativsql/mapper/PropertyMetadata.java
@@ -1,6 +1,6 @@
 package ovh.heraud.nativsql.mapper;
 
-import ovh.heraud.nativsql.db.DatabaseDialect;
+import ovh.heraud.nativsql.db.IdentifierConverter;
 import ovh.heraud.nativsql.util.FieldAccessor;
 import lombok.Getter;
 
@@ -11,15 +11,15 @@ import lombok.Getter;
 public class PropertyMetadata<T> {
     private final FieldAccessor fieldAccessor; // camelCase (e.g., "firstName")
     private final ITypeMapper<T> typeMapper;
-    private final DatabaseDialect dialect;
+    private final IdentifierConverter identifierConverter;
 
-    public PropertyMetadata(FieldAccessor fieldAccessor, ITypeMapper<T> typeMapper, DatabaseDialect dialect) {
+    public PropertyMetadata(FieldAccessor fieldAccessor, ITypeMapper<T> typeMapper, IdentifierConverter identifierConverter) {
         this.fieldAccessor = fieldAccessor;
         this.typeMapper = typeMapper;
-        this.dialect = dialect;
+        this.identifierConverter = identifierConverter;
     }
 
     public String getColumnName() {
-        return dialect.javaToDBIdentifier(fieldAccessor.getName());
+        return identifierConverter.toDB(fieldAccessor.getName());
     }
 }