+++ src/test/java/ovh/heraud/nativsql/repository/mysql/MySQLGroupRepository.java
@@ -1,7 +1,7 @@
 package ovh.heraud.nativsql.repository.mysql;
 
 import ovh.heraud.nativsql.domain.mysql.Group;
-import org.springframework.lang.NonNull;
+import org.jspecify.annotations.NonNull;
 import org.springframework.stereotype.Repository;
 
 /**